refactor: refactored home page hero loading
